requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.45
pyyaml==6.0.1
fuzzywuzzy==0.18.0
python-Levenshtein>=0.25.0
Pillow>=11.0.0
PyPDF2==3.0.1
requests>=2.28.0
python-multipart==0.0.6
aiofiles==23.2.1
feedparser==6.0.10
pdf2image>=1.17.0
bcrypt>=4.1.0
PyJWT>=2.8.0

# Linting and code quality
pylint>=3.0.0
flake8>=6.1.0
black>=24.0.0
isort>=5.13.0